Untar top: If file exits in destination the copy is avoided

+++ tools/file_utils.py
@@ -86,8 +86,10 @@ def untar_top(tar_file, dest_dir=None, top_file=None):
         if not os.path.isdir(dest_dir):
             dest_dir = os.path.dirname(dest_dir)
 
-    shutil.copy2(tar_file, dest_dir)
     new_tar_file = os.path.join(dest_dir, tar_file)
+    if not os.path.exists(new_tar_file):
+        shutil.copy2(tar_file, dest_dir)
+
     with tarfile.open(new_tar_file) as tar:
         tar_name = next(name for name in tar.getnames() if name.endswith(".top"))
         tar.extractall(path=dest_dir)
